+++ src/apify_error.js
@@ -2,10 +2,8 @@ import _ from 'underscore';
 
 export const APIFY_ERROR_NAME = 'ApifyClientError';
 
-export const INVALID_PARAMETER_ERROR_TYPE_V1 = 'INVALID_PARAMETER';
-export const INVALID_PARAMETER_ERROR_TYPE_V2 = 'invalid-parameter';
-export const REQUEST_FAILED_ERROR_TYPE_V1 = 'REQUEST_FAILED';
-export const REQUEST_FAILED_ERROR_TYPE_V2 = 'request-failed';
+export const INVALID_PARAMETER_ERROR_TYPE = 'invalid-parameter';
+export const REQUEST_FAILED_ERROR_TYPE = 'request-failed';
 export const REQUEST_FAILED_ERROR_MESSAGE = 'Server request failed.';
 export const NOT_FOUND_STATUS_CODE = 404;